TCA-870 extract canvas and print renderers to hooks
