+++ src-ts/tools/learn/learn-lib/index.ts
@@ -10,5 +10,7 @@ export * from './learn-level-icon'
 export * from './learn-swr'
 export * from './my-course-card'
 export * from './svgs'
+export * from './use-certificate-canvas-hook'
+export * from './use-certificate-print-hook'
 export * from './use-certificate-scaling-hook'
 export * from './wave-hero'
