+++ src-ts/tools/learn/learn-lib/use-certificate-canvas-hook/index.ts
@@ -1,0 +1,1 @@
+export * from './useCertificateCanvas.hook'
