+++ src-ts/tools/learn/learn-lib/use-certificate-print-hook/index.ts
@@ -1,0 +1,1 @@
+export * from './useCertificatePrint.hook'
